+++ doc/source/history.rst
@@ -6,6 +6,7 @@ History
   - New features:
 
     - :joedb:`tutorial::Memory_Database`
+    - joedbc generates a .gitignore
 
   - Fixes and improvements: